+++ contact/tests/test_views.py
@@ -5,6 +5,7 @@ from bs4 import BeautifulSoup
 from django.urls import reverse
 from django.conf import settings
 
+from conf.tests.test_urls import reload_urlconf
 from contact import forms, views
 
 
@@ -64,7 +65,10 @@ def test_contact_invalid(mock_save, rf):
     'invest-contact',
     'invest-contact-success'
 ))
-def test_contact_pages_localised_urls(url, client):
+def test_contact_pages_localised_urls(url, client, settings):
+    settings.FEATURE_FLAGS['EXPAND_REDIRECT_ON'] = False
+    reload_urlconf(settings)
+
     url = reverse(url) + '?lang=de'
     response = client.get(url)
     soup = BeautifulSoup(response.content, 'html.parser')
@@ -80,7 +84,10 @@ def test_contact_pages_localised_urls(url, client):
     'language_code',
     [code for code, _ in settings.LANGUAGES],
 )
-def test_contact_pages_localised_urls_all_languages(language_code, client):
+def test_contact_pages_localised_urls_all_languages(language_code, client, settings):
+    settings.FEATURE_FLAGS['EXPAND_REDIRECT_ON'] = False
+    reload_urlconf(settings)
+
     url = reverse('invest-contact') + f'?lang={language_code}'
     response = client.get(url)
     soup = BeautifulSoup(response.content, 'html.parser')